jobs_worker/requirements.txt
jstyleson
psycopg2-binary
requests